+++ ProjectTACO/src/com/taco/actor/equipment/MediumWeapon.java
@@ -7,7 +7,7 @@ public class MediumWeapon extends Weapon {
 	public MediumWeapon(Entity owner) {
 		super(owner);
 		damageDoes = 5;
-		delay = (int) (super.delay * 1.2);
+		delay = (short) (super.delay * 1.2);
 		oDelay = delay;
 	}
